reboot02-resume-proxy-api/requirements.txt
flask
fastapi
python-multipart
google-auth
requests
python-dotenv
gunicorn
uvicorn[standard]
httpx[http2]